root : Attempt to fix flashlight by using a symbolic link
OrangeFox tries to enable flashlight by echoing the value of OF_FL_PATH(1/2)/max_brightness to OF_FL_PATH(1/2)/brightness.
A simple way to fix this issue without creating a new node named brighntess for flashlight is to create a symbolic link to the file for enabling flashlight ( /proc/qcom_flash in this case ),
name that symbolic link brightness and create a normal file named max_brightness with the value that enables flashlight.